Reduce testHttpRange dependencies
Initial implementation.
