+++ src/HeaderMangling.h
@@ -6,10 +6,9 @@
  * Please see the COPYING and CONTRIBUTORS files for details.
  */
 
-#ifndef SQUID_SRC_HTTPHEADERTOOLS_H
-#define SQUID_SRC_HTTPHEADERTOOLS_H
+#ifndef SQUID_SRC_HEADERMANGLING_H
+#define SQUID_SRC_HEADERMANGLING_H
 
-#include "acl/forward.h"
 #include "format/Format.h"
 #include "HttpHeader.h"
 #include "sbuf/forward.h"
@@ -23,17 +22,18 @@
 #endif
 
 class HeaderWithAcl;
-class HttpHeader;
-class HttpRequest;
 class StoreEntry;
 
 typedef std::list<HeaderWithAcl> HeaderWithAclList;
+
+#include "acl/forward.h"
 
 /* Distinguish between Request and Reply (for header mangling) */
 typedef enum {
     ROR_REQUEST,
     ROR_REPLY
 } req_or_rep_t;
+
 
 // Currently a POD
 class headerMangler
@@ -117,21 +117,6 @@ public:
     bool quoted;
 };
 
-/// A strtoll(10) wrapper that checks for strtoll() failures and other problems.
-/// XXX: This function is not fully compatible with some HTTP syntax rules.
-/// Just like strtoll(), allows whitespace prefix, a sign, and _any_ suffix.
-/// Requires at least one digit to be present.
-/// Sets "off" and "end" arguments if and only if no problems were found.
-/// \return true if and only if no problems were found.
-bool httpHeaderParseOffset(const char *start, int64_t *offPtr, char **endPtr = nullptr);
-
-bool httpHeaderHasConnDir(const HttpHeader * hdr, const SBuf &directive);
-int httpHeaderParseInt(const char *start, int *val);
-void httpHeaderPutStrf(HttpHeader * hdr, Http::HdrType id, const char *fmt,...) PRINTF_FORMAT_ARG3;
-
-const char *getStringPrefix(const char *str, size_t len);
-
 void httpHdrMangleList(HttpHeader *, HttpRequest *, const AccessLogEntryPointer &al, req_or_rep_t req_or_rep);
 
-#endif /* SQUID_SRC_HTTPHEADERTOOLS_H */
-
+#endif /* SQUID_SRC_HEADERMANGLING_H */
