+++ src/HttpHdrCc.cc
@@ -14,7 +14,8 @@
 #include "HttpHdrCc.h"
 #include "HttpHeader.h"
 #include "HttpHeaderStat.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "sbuf/SBuf.h"
 #include "SquidMath.h"
 #include "StatHist.h"
