+++ src/HttpHdrRange.cc
@@ -12,7 +12,8 @@
 #include "client_side_request.h"
 #include "http/Stream.h"
 #include "HttpHeaderRange.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "Store.h"
 #include "StrList.h"
