+++ src/HttpHdrSc.cc
@@ -15,7 +15,8 @@
 #include "HttpHeader.h"
 #include "HttpHeaderFieldStat.h"
 #include "HttpHeaderStat.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "Store.h"
 #include "StrList.h"
 #include "util.h"
