+++ src/Makefile.am
@@ -259,6 +259,8 @@ squid_SOURCES = \
 	Generic.h \
 	HappyConnOpener.cc \
 	HappyConnOpener.h \
+	HeaderMangling.cc \
+	HeaderMangling.h \
 	HierarchyLogEntry.h \
 	HttpBody.cc \
 	HttpBody.h \
@@ -279,8 +281,6 @@ squid_SOURCES = \
 	HttpHeaderMask.h \
 	HttpHeaderRange.h \
 	HttpHeaderStat.h \
-	HttpHeaderTools.cc \
-	HttpHeaderTools.h \
 	HttpReply.cc \
 	HttpReply.h \
 	HttpRequest.cc \
@@ -1754,239 +1754,23 @@ tests_testHtmlQuote_LDFLAGS = $(LIBADD_DL)
 
 check_PROGRAMS += tests/testHttpRange
 tests_testHttpRange_SOURCES = \
-	$(DELAY_POOL_SOURCE) \
-	$(DNSSOURCE) \
-	$(HTCPSOURCE) \
-	$(IPC_SOURCE) \
-	$(SNMP_SOURCE) \
-	$(UNLINKDSOURCE) \
-	$(WIN32_SOURCE) \
-	AccessLogEntry.cc \
-	AuthReg.h \
-	BodyPipe.cc \
-	tests/stub_CacheDigest.cc \
-	CacheDigest.h \
-	CachePeer.cc \
-	CachePeer.h \
-	CachePeers.cc \
-	CachePeers.h \
-	ClientInfo.h \
-	tests/stub_CollapsedForwarding.cc \
-	ConfigOption.cc \
-	ConfigParser.cc \
-	CpuAffinityMap.cc \
-	CpuAffinityMap.h \
-	CpuAffinitySet.cc \
-	CpuAffinitySet.h \
-	tests/stub_ETag.cc \
-	tests/stub_EventLoop.cc \
-	FadingCounter.cc \
-	FileMap.h \
-	FwdState.cc \
-	FwdState.h \
-	HappyConnOpener.cc \
-	HappyConnOpener.h \
-	HttpBody.cc \
-	HttpBody.h \
-	tests/stub_HttpControlMsg.cc \
-	HttpHdrCc.cc \
-	HttpHdrCc.h \
 	HttpHdrContRange.cc \
 	HttpHdrRange.cc \
-	HttpHdrSc.cc \
-	HttpHdrScTarget.cc \
-	HttpHeader.cc \
-	HttpHeader.h \
-	HttpHeaderFieldStat.h \
-	HttpHeaderTools.cc \
-	HttpHeaderTools.h \
-	tests/testHttpRange.cc \
-	HttpReply.cc \
-	HttpRequest.cc \
-	tests/stub_HttpUpgradeProtocolAccess.cc \
-	tests/stub_Instance.cc \
-	IoStats.h \
-	tests/stub_IpcIoFile.cc \
-	LogTags.cc \
-	MasterXaction.cc \
-	MasterXaction.h \
-	MemBuf.cc \
-	MemObject.cc \
-	tests/stub_MemStore.cc \
-	Notes.cc \
-	Notes.h \
-	Parsing.cc \
-	PeerPoolMgr.cc \
-	PeerPoolMgr.h \
-	Pipeline.cc \
-	Pipeline.h \
-	RefreshPattern.h \
-	RemovalPolicy.cc \
-	RequestFlags.cc \
-	RequestFlags.h \
-	ResolvedPeers.cc \
-	ResolvedPeers.h \
-	SquidMath.cc \
-	SquidMath.h \
-	StatCounters.cc \
-	StatCounters.h \
-	StatHist.cc \
-	StatHist.h \
-	StoreFileSystem.cc \
-	StoreIOState.cc \
-	StoreSwapLogData.cc \
+	http/HeaderTools.cc \
 	StrList.cc \
-	StrList.h \
 	String.cc \
-	Transients.cc \
-	tests/stub_cache_cf.cc \
-	cache_cf.h \
-	cache_manager.cc \
-	tests/stub_carp.cc \
-	carp.h \
-	cbdata.cc \
-	clientStream.cc \
-	tests/stub_client_db.cc \
-	client_side.cc \
-	client_side.h \
-	client_side_reply.cc \
-	client_side_request.cc \
-	dlink.cc \
-	dlink.h \
-	errorpage.cc \
-	event.cc \
-	tests/stub_external_acl.cc \
-	tests/stub_fatal.cc \
-	fatal.h \
-	fd.cc \
-	fd.h \
-	fde.cc \
-	filemap.cc \
-	fqdncache.cc \
-	fqdncache.h \
-	fs_io.cc \
-	fs_io.h \
-	helper.cc \
-	hier_code.h \
-	http.cc \
-	icp_v2.cc \
-	icp_v3.cc \
-	int.cc \
-	int.h \
-	internal.cc \
-	internal.h \
-	tests/stub_ipc_Forwarder.cc \
-	ipcache.cc \
-	tests/stub_libauth.cc \
-	tests/stub_libdiskio.cc \
-	tests/stub_liberror.cc \
-	tests/stub_libeui.cc \
+	tests/stub_cbdata.cc \
+	tests/stub_debug.cc \
+	tests/stub_HttpHeader.cc \
+	tests/stub_HttpReply.cc \
 	tests/stub_libmem.cc \
-	tests/stub_libsecurity.cc \
-	tests/stub_libstore.cc \
-	tests/stub_main_cc.cc \
-	mem_node.cc \
-	mime.cc \
-	mime.h \
-	mime_header.cc \
-	mime_header.h \
-	multicast.cc \
-	multicast.h \
-	neighbors.cc \
-	neighbors.h \
-	pconn.cc \
-	peer_digest.cc \
-	peer_proxy_negotiate_auth.cc \
-	peer_proxy_negotiate_auth.h \
-	peer_select.cc \
-	peer_sourcehash.cc \
-	peer_sourcehash.h \
-	peer_userhash.cc \
-	peer_userhash.h \
-	tests/stub_redirect.cc \
-	redirect.h \
-	refresh.cc \
-	refresh.h \
-	repl_modules.h \
-	stat.cc \
-	stat.h \
-	stmem.cc \
-	store.cc \
-	store_client.cc \
-	tests/stub_store_digest.cc \
-	store_digest.h \
-	store_io.cc \
-	store_key_md5.cc \
-	store_key_md5.h \
-	store_log.cc \
-	store_log.h \
-	store_rebuild.cc \
-	store_rebuild.h \
-	tests/stub_store_stats.cc \
-	store_swapin.cc \
-	store_swapin.h \
-	store_swapout.cc \
-	tools.cc \
-	tools.h \
-	tests/stub_tunnel.cc \
-	tunnel.h \
-	urn.cc \
-	urn.h \
-	tests/stub_wccp2.cc \
-	wccp2.h \
-	wordlist.cc \
-	wordlist.h
-nodist_tests_testHttpRange_SOURCES = \
-	$(BUILT_SOURCES) \
-	tests/stub_libtime.cc
+	tests/stub_MemBuf.cc \
+	tests/testHttpRange.cc
 tests_testHttpRange_LDADD = \
-	libsquid.la \
-	clients/libclients.la \
-	servers/libservers.la \
-	ftp/libftp.la \
-	helper/libhelper.la \
-	http/libhttp.la \
-	parser/libparser.la \
-	acl/libacls.la \
-	acl/libstate.la \
-	acl/libapi.la \
-	proxyp/libproxyp.la \
-	parser/libparser.la \
-	fs/libfs.la \
-	anyp/libanyp.la \
-	icmp/libicmp.la \
-	comm/libcomm.la \
-	ip/libip.la \
-	log/liblog.la \
-	format/libformat.la \
-	$(REPL_OBJS) \
-	$(ADAPTATION_LIBS) \
-	$(SSL_LIBS) \
-	ipc/libipc.la \
-	dns/libdns.la \
-	base/libbase.la \
-	mgr/libmgr.la \
-	html/libhtml.la \
 	sbuf/libsbuf.la \
-	debug/libdebug.la \
-	store/libstore.la \
-	$(SNMP_LIBS) \
-	$(top_builddir)/lib/libmisccontainers.la \
-	$(top_builddir)/lib/libmiscencoding.la \
-	$(top_builddir)/lib/libmiscutil.la \
-	$(LIBCAP_LIBS) \
-	$(LIBGNUTLS_LIBS) \
-	$(LIBHEIMDAL_KRB5_LIBS) \
-	$(REGEXLIB) \
-	$(SSLLIB) \
+	base/libbase.la \
 	$(LIBCPPUNIT_LIBS) \
-	$(LIBSYSTEMD_LIBS) \
 	$(COMPAT_LIB) \
-	$(LIBGSS_LIBS) \
-	$(LIBMIT_KRB5_LIBS) \
-	$(LIBNETFILTER_CONNTRACK_LIBS) \
-	$(LIBNETTLE_LIBS) \
-	$(LIBPSAPI_LIBS) \
 	$(XTRA_LIBS)
 tests_testHttpRange_LDFLAGS = $(LIBADD_DL)
