+++ src/acl/HttpHeaderData.cc
@@ -17,7 +17,8 @@
 #include "cache_cf.h"
 #include "ConfigParser.h"
 #include "debug/Stream.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "sbuf/SBuf.h"
 #include "sbuf/StringConvert.h"
