+++ src/adaptation/icap/ModXact.cc
@@ -27,7 +27,8 @@
 #include "error/Detail.h"
 #include "error/ExceptionErrorDetail.h"
 #include "http/ContentLengthInterpreter.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "MasterXaction.h"
 #include "parser/Tokenizer.h"
