+++ src/auth/basic/Config.cc
@@ -24,7 +24,8 @@
 #include "base64.h"
 #include "cache_cf.h"
 #include "helper.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "mgr/Registration.h"
 #include "rfc1738.h"
