+++ src/auth/digest/UserRequest.cc
@@ -15,7 +15,8 @@
 #include "format/Format.h"
 #include "helper.h"
 #include "helper/Reply.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
 #include "MemBuf.h"
