+++ src/auth/ntlm/Config.cc
@@ -23,7 +23,8 @@
 #include "client_side.h"
 #include "helper.h"
 #include "http/Stream.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
 #include "mgr/Registration.h"
