+++ src/client_side.cc
@@ -94,7 +94,8 @@
 #include "http/one/TeChunkedParser.h"
 #include "http/Stream.h"
 #include "HttpHdrContRange.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
 #include "internal.h"
