+++ src/client_side_reply.cc
@@ -22,7 +22,8 @@
 #include "FwdState.h"
 #include "globals.h"
 #include "http/Stream.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
 #include "ip/QosConfig.h"
