+++ src/errorpage.cc
@@ -23,7 +23,8 @@
 #include "format/Format.h"
 #include "fs_io.h"
 #include "html/Quoting.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "HttpReply.h"
 #include "HttpRequest.h"
 #include "MemBuf.h"
