+++ src/http/ContentLengthInterpreter.cc
@@ -14,7 +14,8 @@
 #include "debug/Stream.h"
 #include "http/ContentLengthInterpreter.h"
 #include "http/one/Parser.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "SquidConfig.h"
 #include "SquidString.h"
 #include "StrList.h"
