+++ src/http/Message.cc
@@ -14,7 +14,8 @@
 #include "http/Message.h"
 #include "http/one/Parser.h"
 #include "HttpHdrCc.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "MemBuf.h"
 #include "mime_header.h"
 #include "SquidConfig.h"
