+++ src/servers/Http1Server.cc
@@ -17,7 +17,8 @@
 #include "comm/Write.h"
 #include "http/one/RequestParser.h"
 #include "http/Stream.h"
-#include "HttpHeaderTools.h"
+#include "HeaderMangling.h"
+#include "http/HeaderTools.h"
 #include "servers/Http1Server.h"
 #include "SquidConfig.h"
 #include "Store.h"
